DB_BACKUP/Mysql/requirements.txt
boto3==1.28.12
botocore==1.31.12
python-dotenv==1.0.0
paramiko==2.11.0
scp==0.14.5